# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: Zawadi — Premium Kenyan Corporate Gifting Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /suppliers:
    get:
      operationId: listSuppliers
      tags: [suppliers]
      summary: List all suppliers
      parameters:
        - name: is_verified
          in: query
          schema: { type: boolean }
        - name: county
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
      responses:
        "200":
          description: List of suppliers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Supplier"
    post:
      operationId: createSupplier
      tags: [suppliers]
      summary: Submit a supplier application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupplierInput"
      responses:
        "201":
          description: Created supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
  /suppliers/{id}:
    get:
      operationId: getSupplier
      tags: [suppliers]
      summary: Get supplier details and story
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Supplier details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
    put:
      operationId: updateSupplier
      tags: [suppliers]
      summary: Update supplier profile
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSupplierInput"
      responses:
        "200":
          description: Updated supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"

  # ── Collections ───────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Supplier:
      type: object
      required: [id, name, county, is_verified, onboarding_status, created_at]
      properties:
        id:
          type: string
        name:
          type: string
        contact_name:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        county:
          type: string
        description:
          type: string
          nullable: true
        story:
          type: string
          nullable: true
        cover_image_url:
          type: string
          nullable: true
        logo_url:
          type: string
          nullable: true
        is_verified:
          type: boolean
        onboarding_status:
          type: string
          enum: [pending, under_review, approved, rejected]
        tags:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
    CreateSupplierInput:
      type: object
      required: [name, county, contact_name, email, phone]
      properties:
        name:
          type: string
        county:
          type: string
        contact_name:
          type: string
        email:
          type: string
        phone:
          type: string
        description:
          type: string
          nullable: true
        story:
          type: string
          nullable: true
    UpdateSupplierInput:
      type: object
      properties:
        name:
          type: string
        county:
          type: string
        contact_name:
          type: string
        email:
          type: string
        phone:
          type: string
        description:
          type: string
          nullable: true
        story:
          type: string
          nullable: true
        cover_image_url:
          type: string
          nullable: true
        logo_url:
          type: string
          nullable: true
        is_verified:
          type: boolean
        onboarding_status:
          type: string
          enum: [pending, under_review, approved, rejected]

    # ── Product ───────────────────────────────────────────────────────────
...